// src/utils/types.ts
export type FormData = {
  name: string;
  email: string;
  message: string;
};
export type KindergartenFormData = {
  contact_type: string;
  kindergarten: string;
  name: string;
  email: string;
  address: string;
  phone: string;
  classroom1: string;
  classroom2: string;
  classroom3: string;
  students: string;
  trial_date: string;
  message: string;
};